- Don't overwrite existing entitlements file - Don't parse the projectfile on AB

+++ src/ios/hooks/install_entitlements.js
@@ -1,4 +1,5 @@
-console.log("Running hook to add iOS Keychain Sharing entitlements");
+// using error to see if this shows up in AB
+console.error("Running hook to add iOS Keychain Sharing entitlements");
 
 var xcode = require('xcode'),
     fs = require('fs'),
@@ -39,47 +40,55 @@ var xcode = require('xcode'),
       throw new Error("Could not find an .xcodeproj folder in: " + iosFolder);
     }
 
-    console.log("Will add iOS Keychain Sharing entitlements to project '" + projName + "'");
+    var destFile = path.join(iosFolder, projName, 'Resources', projName + '.entitlements');
+    if (fs.existsSync(destFile)) {
+      console.error("File exists, not doing anything: " + destFile);
+    } else {
+      console.log("Will add iOS Keychain Sharing entitlements to project '" + projName + "'");
 
-    //var projectPlistPath = path.join(context.opts.projectRoot, 'platforms/ios', projName, util.format('%s-Info.plist', projName));
-    var projectPlistPath = path.join(iosFolder, projName, util.format('%s-Info.plist', projName));
-    var projectPlistJson = plist.parse(fs.readFileSync(projectPlistPath, 'utf8'));
-    var sharedBundleID = projectPlistJson.KeychainSharingBundleID;
-    console.log("KeychainSharingBundleID passed in as variable: " + sharedBundleID);
+      //var projectPlistPath = path.join(context.opts.projectRoot, 'platforms/ios', projName, util.format('%s-Info.plist', projName));
+      var projectPlistPath = path.join(iosFolder, projName, util.format('%s-Info.plist', projName));
+      var projectPlistJson = plist.parse(fs.readFileSync(projectPlistPath, 'utf8'));
+      var sharedBundleID = projectPlistJson.KeychainSharingBundleID;
+      console.log("KeychainSharingBundleID passed in as variable: " + sharedBundleID);
 
-    // create a new entitlements plist file
-    //var sourceFile = 'plugins/cordova-plugin-keychainsharing/src/ios/resources/KeychainSharing.entitlements';
-    var sourceFile = path.join(context.opts.plugin.pluginInfo.dir, 'src/ios/resources/KeychainSharing.entitlements');
-    fs.readFile(sourceFile, 'utf8', function (err, data) {
-      data = data.replace(/__KEYCHAIN_ACCESS_GROUP__/g, sharedBundleID);
-      console.log(data);
 
-      //fs.writeFileSync('platforms/ios/' + projName + '/Resources/' + projName + '.entitlements', data);
-      fs.writeFileSync(path.join(iosFolder, projName, 'Resources',  projName + '.entitlements'), data);
+      // create a new entitlements plist file
+      //var sourceFile = 'plugins/cordova-plugin-keychainsharing/src/ios/resources/KeychainSharing.entitlements';
+      var sourceFile = path.join(context.opts.plugin.pluginInfo.dir, 'src/ios/resources/KeychainSharing.entitlements');
+      fs.readFile(sourceFile, 'utf8', function (err, data) {
+        data = data.replace(/__KEYCHAIN_ACCESS_GROUP__/g, sharedBundleID);
 
-      var projectPath = path.join(projFolder, 'project.pbxproj');
-      var pbxProject = xcode.project(projectPath);
+        fs.writeFileSync(destFile, data);
 
-      // parsing is async, in a different process
-      pbxProject.parseSync();
+        var projectPath = path.join(projFolder, 'project.pbxproj');
 
-      pbxProject.addResourceFile(projName + ".entitlements");
-
-      var configGroups = pbxProject.hash.project.objects['XCBuildConfiguration'];
-      for (var key in configGroups) {
-        var config = configGroups[key];
-        if (config.buildSettings !== undefined) {
-          config.buildSettings.CODE_SIGN_ENTITLEMENTS = '"' + projName + '/Resources/' + projName + '.entitlements"';
-          //console.log("Adding iOS Keychain Sharing entitlements to project '" + projName + "'");
+        var pbxProject;
+        if (context.opts.cordova.project) {
+          pbxProject = context.opts.cordova.project.parseProjectFile(context.opts.projectRoot).xcode;
+        } else {
+          pbxProject = xcode.project(projectPath);
+          pbxProject.parseSync();
         }
-      }
 
-      // write the updated project file
-      fs.writeFileSync(projectPath, pbxProject.writeSync());
-      console.log("OK, added iOS Keychain Sharing entitlements to project '" + projName + "'");
+        pbxProject.addResourceFile(projName + ".entitlements");
 
-      deferral.resolve();
-    });
+        var configGroups = pbxProject.hash.project.objects['XCBuildConfiguration'];
+        for (var key in configGroups) {
+          var config = configGroups[key];
+          if (config.buildSettings !== undefined) {
+            config.buildSettings.CODE_SIGN_ENTITLEMENTS = '"' + projName + '/Resources/' + projName + '.entitlements"';
+            //console.log("Adding iOS Keychain Sharing entitlements to project '" + projName + "'");
+          }
+        }
+
+        // write the updated project file
+        fs.writeFileSync(projectPath, pbxProject.writeSync());
+        console.warn("OK, added iOS Keychain Sharing entitlements to project '" + projName + "'");
+
+        deferral.resolve();
+      });
+    }
   });
 
   return deferral.promise;
